role

+++ Domain/Interfaces/IRolePermissionRepository.cs
@@ -7,6 +7,6 @@ namespace Domain.Interfaces
     public interface IRolePermissionRepository : IGenericRepository<RolePermissionMap>
     {
         IQueryable<RolePermissionMap> GetAllRolePermissionsWithRelatedInfo();
-        Task<RolePermissionMap> GetByIdAsync(int id);
+        Task<RolePermissionMap> GetByIdAsync(long id);
     }
 }
